# backend/artifacts/openapi-v6.0.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClearPath Nexus Engine",
    "version": "6.0.0"
  },
  "paths": {
...
    "/api/v1/predictions/delay": {
      "post": {
        "tags": [
          "machine-learning"
        ],
        "summary": "Predict Delay",
        "operationId": "predict_delay_api_v1_predictions_delay_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DelayFeatureInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelayInferenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/models": {
      "get": {
        "tags": [
          "machine-learning"
        ],
        "summary": "List Models",
        "operationId": "list_models_api_v1_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelSummary"
                  },
                  "type": "array",
                  "title": "Response List Models Api V1 Models Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/models/{model_name}": {
      "get": {
        "tags": [
          "machine-learning"
        ],
        "summary": "Get Model",
        "operationId": "get_model_api_v1_models__model_name__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/models/{model_name}/drift": {
      "get": {
        "tags": [
          "machine-learning"
        ],
        "summary": "Get Model Drift",
        "operationId": "get_model_drift_api_v1_models__model_name__drift_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Name"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 40
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 20,
              "default": 100,
              "title": "Window"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriftReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models/{model_name}/retraining-readiness": {
      "get": {
        "tags": [
          "machine-learning"
        ],
        "summary": "Get Retraining Readiness",
        "operationId": "get_retraining_readiness_api_v1_models__model_name__retraining_readiness_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Name"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 40
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrainingReadiness"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/{prediction_id}/outcome": {
      "post": {
        "tags": [
          "machine-learning"
        ],
        "summary": "Record Prediction Outcome",
        "operationId": "record_prediction_outcome_api_v1_predictions__prediction_id__outcome_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "prediction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Prediction Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionOutcomeInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Record Prediction Outcome Api V1 Predictions  Prediction Id  Outcome Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DelayFeatureInput": {
        "properties": {
          "route_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Id"
          },
          "shipment_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipment Id"
          },
          "route_distance": {
            "type": "number",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Route Distance"
          },
          "segment_count": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Segment Count"
          },
          "departure_hour": {
            "type": "integer",
            "maximum": 23.0,
            "minimum": 0.0,
            "title": "Departure Hour"
          },
          "day_of_week": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 0.0,
            "title": "Day Of Week"
          },
          "month": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Month"
          },
          "current_delay": {
            "type": "number",
            "maximum": 10000.0,
            "minimum": -240.0,
            "title": "Current Delay"
          },
          "journey_progress": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Journey Progress"
          },
          "weather_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Weather Score"
          },
          "rainfall": {
            "type": "number",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Rainfall"
          },
          "temperature": {
            "type": "number",
            "maximum": 80.0,
            "minimum": -80.0,
            "title": "Temperature"
          },
          "wind_speed": {
            "type": "number",
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Wind Speed"
          },
          "visibility": {
            "type": "number",
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Visibility"
          },
          "corridor_congestion": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Corridor Congestion"
          },
          "station_pressure": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Station Pressure"
          },
          "historical_delay": {
            "type": "number",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Historical Delay"
          },
          "schedule_conflicts": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Schedule Conflicts"
          },
          "port_pressure": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Port Pressure"
          },
          "loading_window_delta": {
            "type": "number",
            "maximum": 10000.0,
            "minimum": -10000.0,
            "title": "Loading Window Delta"
          },
          "cargo_weight": {
            "type": "number",
            "maximum": 100000.0,
            "exclusiveMinimum": 0.0,
            "title": "Cargo Weight"
          }
        },
        "type": "object",
        "required": [
          "route_distance",
          "segment_count",
          "departure_hour",
          "day_of_week",
          "month",
          "current_delay",
          "journey_progress",
          "weather_score",
          "rainfall",
          "temperature",
          "wind_speed",
          "visibility",
          "corridor_congestion",
          "station_pressure",
          "historical_delay",
          "schedule_conflicts",
          "port_pressure",
          "loading_window_delta",
          "cargo_weight"
        ],
        "title": "DelayFeatureInput"
      },
      "DelayInferenceResponse": {
        "properties": {
          "prediction_type": {
            "type": "string",
            "title": "Prediction Type"
          },
          "operational_prediction_minutes": {
            "type": "number",
            "title": "Operational Prediction Minutes"
          },
          "deterministic_prediction_minutes": {
            "type": "number",
            "title": "Deterministic Prediction Minutes"
          },
          "ml_prediction_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ml Prediction Minutes"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "model_status": {
            "type": "string",
            "title": "Model Status"
          },
          "used_for_operational_decision": {
            "type": "boolean",
            "title": "Used For Operational Decision"
          },
          "fallback_used": {
            "type": "boolean",
            "title": "Fallback Used"
          },
          "fallback_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason"
          },
          "artifact_checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Checksum"
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms"
          },
          "prediction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Prediction Id"
          },
          "provenance_record_id": {
            "type": "string",
            "format": "uuid",
            "title": "Provenance Record Id"
          },
          "feature_schema_version": {
            "type": "string",
            "title": "Feature Schema Version"
          },
          "confidence_interval": {
            "$ref": "#/components/schemas/PredictionInterval"
          },
          "explanation": {
            "$ref": "#/components/schemas/PredictionExplanation"
          }
        },
        "type": "object",
        "required": [
          "prediction_type",
          "operational_prediction_minutes",
          "deterministic_prediction_minutes",
          "ml_prediction_minutes",
          "model_name",
          "model_version",
          "model_status",
          "used_for_operational_decision",
          "fallback_used",
          "fallback_reason",
          "artifact_checksum",
          "latency_ms",
          "prediction_id",
          "provenance_record_id",
          "feature_schema_version",
          "confidence_interval",
          "explanation"
        ],
        "title": "DelayInferenceResponse"
      },
...
      "DriftReport": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "status": {
            "type": "string",
            "enum": [
              "INSUFFICIENT_DATA",
              "STABLE",
              "WATCH",
              "DRIFT_DETECTED"
            ],
            "title": "Status"
          },
          "labeled_sample_count": {
            "type": "integer",
            "title": "Labeled Sample Count"
          },
          "minimum_sample_count": {
            "type": "integer",
            "title": "Minimum Sample Count"
          },
          "observed_mae_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed Mae Minutes"
          },
          "observed_bias_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed Bias Minutes"
          },
          "reference_mae_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Mae Minutes"
          },
          "mae_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mae Ratio"
          },
          "window_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Start"
          },
          "window_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window End"
          },
          "operational_mode_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Operational Mode Counts"
          },
          "source_line": {
            "additionalProperties": true,
            "type": "object",
            "title": "Source Line"
          }
        },
        "type": "object",
        "required": [
          "model_name",
          "model_version",
          "status",
          "labeled_sample_count",
          "minimum_sample_count",
          "source_line"
        ],
        "title": "DriftReport"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelSummary": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "production": {
            "type": "boolean",
            "title": "Production"
          },
          "algorithm": {
            "type": "string",
            "title": "Algorithm"
          },
          "artifact_checksum": {
            "type": "string",
            "title": "Artifact Checksum"
          },
          "feature_schema_version": {
            "type": "string",
            "title": "Feature Schema Version"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          },
          "eligibility": {
            "additionalProperties": true,
            "type": "object",
            "title": "Eligibility"
          }
        },
        "type": "object",
        "required": [
          "model_name",
          "version",
          "status",
          "production",
          "algorithm",
          "artifact_checksum",
          "feature_schema_version"
        ],
        "title": "ModelSummary"
      },
...
      "PredictionInterval": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "lower_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lower Minutes"
          },
          "upper_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upper Minutes"
          },
          "nominal_coverage": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Nominal Coverage"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "calibration_rows": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Calibration Rows"
          },
          "calibration_source_types": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Calibration Source Types"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "available",
          "method",
          "calibration_rows"
        ],
        "title": "PredictionInterval"
      },
      "PredictionOutcomeInput": {
        "properties": {
          "actual_remaining_delay_minutes": {
            "type": "number",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Actual Remaining Delay Minutes"
          },
          "data_class": {
            "type": "string",
            "const": "OPERATOR_CONFIRMED",
            "title": "Data Class",
            "default": "OPERATOR_CONFIRMED"
          }
        },
        "type": "object",
        "required": [
          "actual_remaining_delay_minutes"
        ],
        "title": "PredictionOutcomeInput"
      },
...
      "RetrainingReadiness": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_READY",
              "NOT_NEEDED",
              "READY_FOR_MANUAL_REVIEW"
            ],
            "title": "Status"
          },
          "gates": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Gates"
          },
          "evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Evidence"
          },
          "manual_approval_required": {
            "type": "boolean",
            "const": true,
            "title": "Manual Approval Required",
            "default": true
          },
          "automatic_retraining": {
            "type": "boolean",
            "const": false,
            "title": "Automatic Retraining",
            "default": false
          },
          "automatic_promotion": {
            "type": "boolean",
            "const": false,
            "title": "Automatic Promotion",
            "default": false
          },
          "recommended_action": {
            "type": "string",
            "title": "Recommended Action"
          },
          "source_line": {
            "additionalProperties": true,
            "type": "object",
            "title": "Source Line"
          }
        },
        "type": "object",
        "required": [
          "model_name",
          "model_version",
          "status",
          "gates",
          "evidence",
          "recommended_action",
          "source_line"
        ],
        "title": "RetrainingReadiness"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}